submitting now works
Also, changed the way that page uis are added to the page -
the left/right/top/bottom position values are all set so that
the page completely fills its container.  That way, we don't have
to set a size (width/height) in any of the page UI classes,
but they will still expand to fill the available space.