added ISubmitService interface, implementation
The initial implementation is "Out Of Process",
meaning that external build servers connect and
do the actual work of compiling and testing
submissions.  The SubmitServiceImpl servlet
simply forwards submit requests to the
OutOfProcessSubmitService singleton, which then
puts it on a queue.  When a worker thread
representing an idle build server grabs it from
the queue, it then compiles/tests it and
makes the TestResults available (so that they can
be sent back to the client.)

The implementation currently does not do any
actual compiling or testing.