Cleaner support for compiling and testing.  Refactored code so that a Tester takes Submission and produces a SubmissionResult.
Really horrible classloading issue where TestResult cannot be loaded in
the KillableTaskManager if it hasn't already been loaded.  This means
that the code in JavaTester, which uses its own security
manager, needs to be checked because the security manager and the
classloader may be interacting in funny ways.  Could be loading two
copies of TestResult with two different classloaders.  Ugh.