Major re-factoring:  Submissions are sent to the Builder, and it returns SubmissionResult.
CompilationResult will be inside a SubmissionResult, and
CompilationResult contains a list of CompilerDiagnostics, which contain
the compiler error messages.  Server needs to display those error
messages in some way to indicate that the code doesn't compile.

We could really use hotlinks to problems (either breadcrumbs or request
params that go straight to a certain problem, probably using the
problemID).