change ProcessRunner to use exit status file made by runProcess.pl
Also, fixed a bug in runProcess.pl.

The exit status (such as whether or not the process was terminated
by a signal) does seem to be detected correctly.  Need to adjust
the methods that create TestOutcomes to make use of this information
(so the student will know if the program crashed for a particular
test case.)