change DevelopmentPage to use SplitLayoutPanel
That way, the user can adjust the relative sizes of the
different sub-windows.

Had to change most of the views to extend ResizeComposite rather
than Composite, so that child widgets get resized correctly
when the sliders are moved.  Also had to change AceEditor to
implement RequiresResize.

Unrelated change: changed Ace to use Solarized Light theme.
Not sure if I like it, although it's less garish than
Vibrant Ink.