store activity on server, restore it if user leaves page and comes back
The "activity" is the UI page that the user was working with,
and the initial session objects for that page.  This functionality
handles the case where the user navigates away from the CloudCoder
webpage and then comes back; as long as their server session is
still active, we can bring them back to what they are working on.