eliminated serialization warnings on compilation
Created a new interface ActivityObject which extends Serializable.
Any session objects that should be saved in the Activity
(so they can be restored on page reload) should implement this
interface.  Trying to put all Serializable session objects
in the Activity confuses gwtc because the set of possible types
is too large.