factored out IProblem and IProblemData from Problem and ProblemData
These interfaces will enable the use of adapter objects with
Problem/ProblemData objects.  For example, an adapter could act as
a publisher for state changes made to a Problem object (which
would be useful in a UI for editing problems.)  Implementing
publish/subscribe in Problem directly would be undesirable since
Problems are communicated by GWT serialization, are persisted,
etc.