hack for Jetty not allowing overriding of servlet ctx init parms
Technically, Jetty does allow overriding servlet context init
params, but it requires the use of an external override file,
which is in XML.  Why there isn't a programmatic interface to
do this is anybody's guess.

Since the communication we need is between CloudCoderDaemon
(initializing the webapp) and JDBCDatabaseConfigServletContextListener,
added a specific hack to both that stuffs the needed config
properties in a system property.