fixed bug: changes to editors in EditProblemPage weren't taking effect
The problem is that commit() was setting a field of the
ProblemAdapter, which was then turning around and calling update()
on all of the editors, making them discard their current contents
in favor of the (stale) data in the model object.  Fixed this by
adding an onModelObjectChange() method to EditModelObjectField,
which subclasses can override in the case where they need to
respond to changes in a field they are NOT editing.

This bug was making me seriously doubt my sanity!