implement a meta-model for persistable model objects
The meta-model describes the names, types, and other properties
of model object fields, and is used by the persistence layer to
store and retrieve model objects.

Added meta-information to Problem and ProblemData.  Used it in
JDBCDatabase to generate an update statement to update a problem.

Eventually, the database itself can be created from this
meta-information, meaning that we don't have to worry about the
model objects getting out of sync with the database schema.
This will also allow validation of things like field lengths,
since they will be explictly represented as part of the model object
classes.