use override-web.xml to override ctx params, build script improvements
Jetty allows overriding of web.xml context params by specifying
an additional web.xml whose values override the one in the
webapp's actual web.xml.  We now use this mechanism to allow
the configuration properties in local.properties (generated by
configure.pl) to override the ones in web.xml, which are suitable
only for development.

Also, improvements to build.xml scripts.  Specifically,
we don't copy cloudcoderModelClasses.jar and
cloudcoderModelClassesPersist.jar into the webapp's
WEB-INF/lib directory any more, but instead resolve build-time
and development-time dependencies on them through the
classpath.  CloudCoderWebServer's build.xml copies these
jar files into WEB-INF/lib as part of creating
cloudcoderApp.jar, so they are never in the actual
filesystem.  This is important, because a stale version
of either jar file at development time can lead to strange
behavior, as the source code can get out of sync with the
class files (in the jars) that the webapp is actually
using.

Also, modified the build.xmls so that needed libraries and
code are built automatically, for proper prerequisite
management.