allow CreateWebappDatabase to run from deployable jarfile
This will simplify deployment, and gives us a nice way to
allow the user to customize their installation, because
CreateWebappDatabase can prompt the user for values
such as the institution name.  In the future, we could also
allow other administrative kinds of tasks to be done
by the deployable jarfile.