IProblem and IProblemData now have their own schemata
This allows Problem and ProblemData to define their own schemata
which are parameterized with the correct type (e.g.,
ModelObjectSchema<Problem> rather than ModelObjectSchema<IProblem>.)

This should allow an IModelObject interface with a getSchema()
method to be defined, which will be useful for implementing
generic operations (such as XML serialization and persistence)
which access a model object's schema without requiring that schema
to be passed as a separate parameter.  This should result
(eventually) in significantly simpler persistence and XML
serialization code.