move schemas from IProblem{Data} to Problem{Data}; other schema changes
Other changes:

Make it easier to create an updated subclass schema that is
based on a previous version of the subclass schema with
deltas from a superclass schema applied.  This is necessary
when a field is added to a superclass and must be reflected in
the superclass schema and the schemas of all subclasses.

Renamed ModelObjectSchema.deltaFrom() to ModelObject.basedOn():
this seems more readable.