delete problems easily and safely by just setting a 'deleted' flag
Physically deleting a problem record is dangerous because
lots of other records are linked to it, and if a problem id is
reused then suddenly stale/invalid data is linked to the new
problem.  Setting a deleted flag just means that searches for
the problem need to ignore deleted problems.