use dialog to log back in if an RPC call fails due to session timeout
I only added support for this in a couple places - need to do
more work.  Pretty much anywhere there is an onFailure() async
callback method is a potential site for use.

The SessionExpiredDialogBox class needs some work - I can't figure
out how to get an error message to be visible if initially blank.