use dynamically-typed artifacts for problem/test cases/program text
Also, added a ProgramSource class so that the program text isn't
just a plain string.

Modified BuilderSubmission so that artifacts are kept in a map
indexed by artifact type (class).  This allows artifacts to be
replaced easily, allowing build steps to replace an artifact
(meaning that arbitrary transformations of artifacts are possible).