created CommandInput class to represent input to a Command
Also, added CreateCommandInputsForEachTestCaseBuildStep, to
create a CommandInput for each TestCase.

Added toCommand method to NativeExecutable, to create a
Command to run the executable with arbitrary arguments.

Changed command-running build step to use an array of Command
objects (with corresponding CommandInputs), so that the
test executable can be executed with different arguments.