initial support for RUBY_METHOD submissions - not working yet
One big problem we have is that the ThreadGroupSecurityManager
disallows permissions that JRuby needs.  Need to figure out a
way to have different SecurityManagers for different testers.

As far as testing the Ruby code, it almost works (generated test
code works fine in CRuby), but for some reason the call to the
test scaffold method (_test) fails in JRuby.