script to fetch all required jar files
Keeping all of the jarfiles checked into the
repository is eventually going to cause the repository
to be too large.  As long as we keep this script up
to date, and assuming that the locations where the
jarfiles are archived are stable, this should be a
reasonable solution.

Right now, only the jarfiles in CloudCoderLogging
are fetched.  They are not identical to the ones we
have checked in, but I think that's because the
script downloads them from maven repositories
(where I think some extra metadata is added.)