more work on modules class/table - not quite working yet
Added a delta type for persisting a model object, so that
a migration will create the default "uncategorized" module.

What is still needed is a way to specify a default value for
an added column.  The default module id in problems will need
to be 1, because mysql treats "0" in an insert into an
auto increment column as a request to auto-generate an id, not
insert a literal 0. Fucking mysql.  So, we'll need the id
of the default module to be 1, not 0.