hacking on code to get problems/subscription receipts for user
Trying to allow an instructor to retrieve all problems and
best subscription receipts for another user.

Doesn't quite work yet because the code assumes that the user
is requesting his/her OWN subscription receipts.  The database
query needs to be modified to allow the query if the requesting
user is an instructor, but otherwise allow the requesting
user to see problems that he/she would otherwise be
entitled to see.