support getting text for user's first submission receipt
The user's first submission receipt will have its status set to
SubmissionStatus.STARTED, and does not correspond to any
full text change.  So, if the database is asked to get the
submission text for this kind of submission receipt, retrieve
the first full-text change, which will typically be the
problem skeleton.