for load tester, make deep copy of each EditSequence
LoadTesterTask (and the PlayEditSequence object it uses) need to
modify the Change objects in the EditSequence to match the
user account.  If we don't make a deep copy of the EditSequence,
many tasks will share the EditSequence and their changes to the
user id of the Change/Event objects will conflict with each
other, leading to problems when the tasks try to send Changes
(edits) to the webapp.