random delay prior to initial login; retry RPC ops
The random delay prior to login is to avoid large numbers
of simulated clients attempting to log in at exactly the same
time (which can lead to RPC timeouts).

The RPC operation retry is to simulate what students
would likely do on a failed login: try again.  The
LoadTesterActivityReporter keeps track of recoverable
and unrecoverable RPC exceptions.