working on supporting direct links to exercises
Allow both a "fragment name" selecting a particular page,
and "page params", meaning a description of how to instantiate
session objects, to be specified as part of the URL.
Initially, we will try to allow direct links to an exercise
by providing course id and problem id as part of the page
params for the DevelopmentPage.

Does work yet, but some of the infrastructure is in place.