more debugging of URL fragments: not working
Removed GWT history support, in case that was interfering.

I'm getting completely screwy behavior: the initial fragment
is invisible to onModuleLoad, and instead it sees the fragment
name that is set LATER when a CloudCoderPage is activated.
Extremely weird, it's like onModuleLoad predicts the future
somehow, or it's getting executed twice (but I can only
debug the second time.)