found and fixed mysterious changing of the URL fragment name
The first time we attempted to change the URL fragment,
when going to the LoginPage, it was triggering a page
reload in the browser because the ":" (part of the
GWT development mode URL) was getting URL-encoded.
Undoing that encoding left the non-fragment part of the
URL unchanged, allowing the URL change to take place
without a page reload.

Yay?