added --check option to fetchdeps.pl
This checks for any "unexpected" jarfiles - ones not
explicitly downloaded by fetchdeps.pl or explicitly
mentioned in a .gitignore file.

The presence of extraneous jarfiles can cause serious
issues in build artifacts, since the wrong versions
of classes could be included.  (This happened with a
stale version of daemon.jar that caused a deployed
builder to function incorrectly.)