added a general mechanism for loading page objects
The PageLoader class can load any needed page objects
based on PageParams specified as part of the URL fragment.

Removed all of the special-purpose code in DevelopmentPage
for loading CourseSelection and Problem from parameters in
the URL fragment.

Still to do: have all of the CloudCoderPages other than
DevelopmentPage explicitly declare which objects they need,
and implement support in LoadPageObjects to load those kinds
of objects.