implemented error UI for when page objects can't be loaded
This is shown in the case where cloudcoder is reached
via a direct link to a page, but the page objects specified
by the page params can't be loaded, or when some other error
prevents a directly-linked page from loading properly.

The error UI allows the user to navigate to the home page
(CoursesAndProblemsPage2).