pass SIGTERM through wrapper script to actual test process
If the builder wants to forcibly stop a test process, then
it calls killProcess on the ProcessRunner, which in turn
calls Process.destroy on the test subprocess, which sends
SIGTERM to the subprocess.  HOWEVER, the immediate subprocess
known to Java is actually the wrapper script, so all this
accomplishes is killing the wrapper script, leaving the real
test process running!

Created a new version of the wrapper script, runProcess2.sh.
It catches SIGTERM and delivers it to the real subprocess.