new layout: put problem description view and editor side-by-side
This is a WAY better use of space, and should allow
students to see the problem description more easily.

This isn't quite ready: the north panel (with the problem
name and brief description and the page nav panel)
needs to be a fixed height, and should not have a slider.
I think we can fix this by not having the top-level SplitLayoutPanel
manage it, and simply not having a north panel.