more progress on health monitor: almost working
Changed the Health servlet so that there is a proper model
object to represent webapp instance health, convertable
to/from JSON.

Implemented the actual checking and status reporting
in HealthMonitor.

Added a TestMain to allow testing.

Current failure mode is that it reports that there are no
builder threads (even though the actual instances being
checked to report having a positive number of connected
builder threads.)  Need to investigate.