if not using SSL, allow connections only from localhost and host's IP
This is kind of cheesy: the webapp is still listening for
connections from all sources.  It would be nice if there were
a way that a single server socket could listen on a specified
SET of addresses.  (Actually: probably JUST listening on the
host IP as determined from the cloudcoder.submitsvc.oop.host
property might work.)