add superuser flag to User class
This is a fairly large change, because I needed to add
support for database migrations which modify existing
model objects: the "bootstrap" user (user id=1) is automatically
upgraded to a superuser.  This required changing
ModelObjectSchema to store a reference to the class object
of the type described by the schema, so that the migration
can instantiate an object to carry out the transformation on.

Also updated CreateWebappDatabase to automatically set the
superuser flag in the initial user account.