fix bugs in EditProblemPage, EditStringFieldWithAceEditor
EditProblemPage was setting the TestCase on the TestCaseEditors
prior to having added them to the DOM, which in general is
a bad idea, since this involves a call to update() which
could access UI elements which haven't been initialized yet.

EditStringFieldWithAceEditor wasn't setting any default
mode and theme, so set reasonable defaults.