added ModuleListBox, fixed subtle bug in SessionUtil RPC
The RPC bug was that the recursive restart of an RPC
call following a successful reauthentication called the
top-level SessionUtil RPC method, which due to
RunOnce/OneTimeRunnable would already consider there to
be an RPC call in progress.  Fix is to create a properly
recursive method to do the actual RPC within each
OneTimeRunnable object.