bug workaround: problem list not updated when in inactive tab
The problem seems to be that if the Exercises tab isn't
selected, and the course selection is changed, the problem
list isn't updated.  (The code to do the update does executed,
but the actual UI isn't updated.)

As a workaround eagerly refresh the UI when the Exercises
tab is selected.