refactoring of course instructor UIs, start ManageUsersPanel
ManageUsersPanel will basically be a version of the current
UserAdminPage UI, but implementing only the functionality
that isn't already available from the other Manage course
tab UIs.

The refactoring eliminates CourseInstructorFormUI in favor
of a more general approach to monitoring and notification of
changes in instructor status that don't require the use of
a specific superclass.  This is motivated by ManageUsersPanel
not being a form UI.