added dockerrun.pl entry point script
The entry point script starts mysqld and the
CloudCoder webapp, waits for SIGTERM, and then shuts
down the CloudCoder webapp and mysqld cleanly.

At this point we seem to be getting fairly close
to having a useful containerized CloudCoder webapp.
More work needs to be done to allow post-deployment
customization.