defer keystore creation/configuration
We can't create the keystore as part of Docker image
creation, otherwise every CloudCoder instance running
from the Docker image would have the same keystore.
Changed bootstrap.pl to defer keystore creation and
configuration, and support it as a runtime step.
Changed dockerrun.pl to check whether there is a
real keystore, and if not, generate and configure
one.

Not entirely tested yet.