minor tweaks to deferred keystore generation
Bootstrap script now checks to see whether mysql
is running, and only attempts to start it if
it's actually necessary.

The docker entry point script prints the discovered
keystore size (just to confirm that we're reliably
detecting whether or not the keystore needs to
be generated/configured.)

These features are tested and appear to be working
correctly at this point.