added support for an integrated builder
The integrated builder runs with a single thread and
as a separate user.  The idea is that it guarantees
that a builder is always running.  Eventually the
goal is to spin up additional external builders when
the submission queue grows past a specified threshold.

Also, fixed bugs with creation of the global
CLOUDCODER_VERSION file (needs to be sudoed.)