make sure that webapp can resolve its own hostname
There is some ill-advised code in the submission queue's
ServerTask that tries to prevent connections from external
hosts if SSL is not being used.  This has the
unfortunate side-effect that if the webapp server's
hostname is not currently resolvable by DNS, it won't
accept *any* builder connections.

An easy fix is to just set this to localhost for the
webapp's cloudcoder.properties.