get best submissions before arbitrary deadline
This will be useful for checking whether students have completed
the required CloudCoder exercises before the beginning of
class (with each section having a different class time.)

The doGetBestSubmissionReceipts method in Queries, and the
various methods/classes which use it, are modified to require
a maxTs (max timestamp) value.  Submissions later than this
max timestamp are not considered.

Also: replaced the mysql connector (which was very old)
with the most recent mariadb connector.  This actually
triggered a bug in NestedJarClassLoader, which is that
jar entries aren't guaranteed to specify a size (which
is true of the mariadb connector jar file), and for such
entries we can't preallocate the byte buffer.  Doing this
dynamically using a ByteArrayOutputStream is just as easy.