major refactoring of ProgSnap2 exporter
Use a single database query to retrieve all of the events
for a student/problem, "in order" (at least the the extent
that the database can reconstruct the order.)  Using
left outer joins, the query retrieves associated Change
and SubmissionReceipt objects.  Doing the retrieval
this way should eliminate the need for post-processing steps,
e.g., to determine code state ids for non-edit events.